保存信息

+++ yn-core/src/main/java/com/yn/model/Wallet.java
@@ -3,6 +3,7 @@ package com.yn.model;
 import com.yn.domain.IDomain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -16,7 +17,7 @@ public class Wallet extends IDomain implements Serializable {
 	@Column(columnDefinition = "int(11) comment '[用户id]'")
 	private Long userId;
 	@Column(insertable = false, precision = 12, scale = 2, columnDefinition = "decimal(12,2) default 0 comment '[余额 元]'")
-	private Double money;
+	private BigDecimal money;
 	@Column(insertable = false, precision = 12, scale = 2, columnDefinition = "decimal(12,2) default 0 comment '[积分 元]'")
 	private Double integral;
 	
@@ -27,10 +28,11 @@ public class Wallet extends IDomain implements Serializable {
 	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public Double getMoney() {
+	
+	public BigDecimal getMoney() {
 		return money;
 	}
-	public void setMoney(Double money) {
+	public void setMoney(BigDecimal money) {
 		this.money = money;
 	}
 	public Double getIntegral() {
